Ref #4929 - Migration: Project templates must be migrated before release
 * use ms-appx-web:// for Windows platoforms
 * replace some '<gap:' to '<' for offchat and retailstore (fix the splash missing on local build)

+++ Studio/src/com/twinsoft/convertigo/engine/admin/services/mobiles/MobileResourceHelper.java
@@ -171,8 +171,10 @@ public class MobileResourceHelper {
 								if (inFile.exists()) {
 									file = file.replace("scripts/", "js/");
 									File outFile = new File(destDir, file);
-									outFile.getParentFile().mkdirs();
-									FileUtils.copyFile(inFile, outFile, true);
+									if (!outFile.exists()) {
+										outFile.getParentFile().mkdirs();
+										FileUtils.copyFile(inFile, outFile, true);
+									}
 									line = line.replaceFirst("\"\\.\\./\\.\\./\\.\\./\\.\\./.*?\"", "\"" + file + "\"");
 									
 									handleJQMcssFolder(inFile, outFile);
